requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
python-decouple==3.8
django-cors-headers==4.3.1
drf-spectacular==0.27.0
amqp==5.2.0
django-filter==23.5
